+++ packages/api-client/tests/test-executions.js
@@ -275,3 +275,65 @@ test('bulkDeleteByCollection calls the callback with the expected object and ret
     callback,
   }));
 });
+
+test('bulkArchiveExecutions calls the callback with the expected object and returns the parsed response', async (t) => {
+  const body = {
+    batchSize: 100,
+    expirationDays: 200,
+  };
+
+  const expected = {
+    prefix: t.context.testPrefix,
+    payload: {
+      httpMethod: 'PATCH',
+      resource: '/{proxy+}',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      path: '/executions/archive/',
+      body: JSON.stringify(body),
+    },
+    expectedStatusCodes: 200,
+  };
+
+  const callback = (configObject) => {
+    t.deepEqual(expected, configObject);
+  };
+
+  await t.notThrowsAsync(executionsApi.bulkArchiveExecutions({
+    prefix: t.context.testPrefix,
+    body,
+    callback,
+  }));
+});
+
+test('bulkArchiveExecutionsAsync calls the callback with the expected object and returns the parsed response', async (t) => {
+  const body = {
+    batchSize: 100,
+    expirationDays: 200,
+  };
+
+  const expected = {
+    prefix: t.context.testPrefix,
+    payload: {
+      httpMethod: 'PATCH',
+      resource: '/{proxy+}',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      path: '/executions/archiveAsync/',
+      body: JSON.stringify(body),
+    },
+    expectedStatusCodes: 200,
+  };
+
+  const callback = (configObject) => {
+    t.deepEqual(expected, configObject);
+  };
+
+  await t.notThrowsAsync(executionsApi.bulkArchiveExecutionsAsync({
+    prefix: t.context.testPrefix,
+    body,
+    callback,
+  }));
+});