+++ packages/api/endpoints/executions.js
@@ -3,6 +3,7 @@
 'use strict';
 
 const router = require('express-promise-router')();
+const moment = require('moment');
 const { v4: uuidv4 } = require('uuid');
 const { z } = require('zod');
 const isError = require('lodash/isError');
@@ -18,6 +19,7 @@ const {
   ExecutionPgModel,
   translatePostgresExecutionToApiExecution,
   ExecutionSearch,
+  TableNames,
 } = require('@cumulus/db');
 const { deconstructCollectionId } = require('@cumulus/message/Collections');
 
@@ -224,6 +226,36 @@ async function del(req, res) {
   return res.send({ message: 'Record deleted' });
 }
 
+const bulkArchiveExecutionsSchema = z.object({
+  batchSize: z.number().positive().optional().default(100),
+  expirationDays: z.number().positive().optional().default(365),
+});
+const parsebulkArchiveExecutionsPayload = zodParser('bulkArchiveExecutions payload', bulkArchiveExecutionsSchema);
+async function bulkArchiveExecutions(req, res) {
+  const {
+    getKnexClientMethod = getKnexClient,
+  } = req.testContext || {};
+  const body = parsebulkArchiveExecutionsPayload(req.body);
+  if (isError(body)) {
+    return returnCustomValidationErrors(res, body);
+  }
+  const knex = await getKnexClientMethod();
+  const expirationDate = moment().subtract(body.expirationDays, 'd').format('YYYY-MM-DD');
+
+  const subQuery = knex(TableNames.executions)
+    .select('cumulus_id')
+    .whereBetween('updated_at', [
+      new Date(0),
+      expirationDate,
+    ])
+    .where('archived', false)
+    .limit(body.batchSize);
+  const updatedCount = await knex(TableNames.executions)
+    .update({ archived: true })
+    .whereIn('cumulus_id', subQuery);
+  return res.send({ recordsUpdated: updatedCount });
+}
+
 /**
  * Get execution history for a single granule or multiple granules
  *
@@ -359,6 +391,39 @@ async function bulkDeleteExecutionsByCollection(req, res) {
   return res.status(202).send({ id: asyncOperationId });
 }
 
+const bulkArchiveExecutionsAsyncWrapperSchema = z.object({
+  batchSize: z.number().optional().default(10000),
+  expirationDays: z.number().optional().default(365),
+});
+const parseBulkArchiveExecutionsAsyncWrapperPayload = zodParser('bulkChangeCollection payload', bulkArchiveExecutionsAsyncWrapperSchema);
+
+async function bulkArchiveExecutionsAsyncWrapper(req, res) {
+  const payload = parseBulkArchiveExecutionsAsyncWrapperPayload(req.body);
+  const asyncOperationId = uuidv4();
+  const asyncOperationEvent = {
+    asyncOperationId,
+    callerLambdaName: getFunctionNameFromRequestContext(req),
+    lambdaName: process.env.ArchiveRecordsLambda,
+    description: 'look at me go!',
+    operationType: 'Bulk Granule Reingest',
+    payload: {
+      config: {
+        ...payload,
+        recordType: 'executions',
+      },
+    },
+  };
+  log.debug(
+    `About to invoke lambda to start async operation ${asyncOperationId}`
+  );
+  await startAsyncOperation.invokeStartAsyncOperationLambda(
+    asyncOperationEvent
+  );
+  return res.status(202).send({ id: asyncOperationId });
+}
+
+router.patch('/archiveAsync', bulkArchiveExecutionsAsyncWrapper);
+router.patch('/archive', bulkArchiveExecutions);
 router.post('/search-by-granules', validateGranuleExecutionRequest, searchByGranules);
 router.post('/workflows-by-granules', validateGranuleExecutionRequest, workflowsByGranules);
 router.post(
@@ -375,5 +440,6 @@ router.delete('/:arn', del);
 module.exports = {
   del,
   router,
+  bulkArchiveExecutions,
   bulkDeleteExecutionsByCollection,
 };