+++ packages/api/endpoints/granules.js
@@ -5,7 +5,7 @@
 const { z } = require('zod');
 const isError = require('lodash/isError');
 const pMap = require('p-map');
-
+const moment = require('moment');
 const router = require('express-promise-router')();
 const cloneDeep = require('lodash/cloneDeep');
 const { v4: uuidv4 } = require('uuid');
@@ -16,7 +16,7 @@ const {
 const Logger = require('@cumulus/logger');
 const { constructCollectionId, deconstructCollectionId } = require('@cumulus/message/Collections');
 const { RecordDoesNotExist } = require('@cumulus/errors');
-const { GranuleSearch } = require('@cumulus/db');
+const { GranuleSearch, TableNames } = require('@cumulus/db');
 
 const { ExecutionAlreadyExists } = require('@cumulus/aws-client/StepFunctions');
 
@@ -823,6 +823,35 @@ async function bulkPatch(req, res) {
   });
 }
 
+const bulkArchiveGranulesSchema = z.object({
+  batchSize: z.number().positive().optional().default(100),
+  expirationDays: z.number().positive().optional().default(365),
+});
+const parsebulkArchiveGranulesPayload = zodParser('bulkArchiveGranules payload', bulkArchiveGranulesSchema);
+async function bulkArchiveGranules(req, res) {
+  const {
+    getKnexClientMethod = getKnexClient,
+  } = req.testContext || {};
+  const body = parsebulkArchiveGranulesPayload(req.body);
+  if (isError(body)) {
+    return returnCustomValidationErrors(res, body);
+  }
+  const knex = await getKnexClientMethod();
+  const expirationDate = moment().subtract(body.expirationDays, 'd').format('YYYY-MM-DD');
+
+  const subQuery = knex(TableNames.granules)
+    .select('cumulus_id')
+    .whereBetween('updated_at', [
+      new Date(0),
+      expirationDate,
+    ])
+    .where('archived', false)
+    .limit(body.batchSize);
+  const updatedCount = await knex(TableNames.granules)
+    .update({ archived: true })
+    .whereIn('cumulus_id', subQuery);
+  return res.send({ recordsUpdated: updatedCount });
+}
 /**
  * Delete a granule by granuleId
  *
@@ -1328,8 +1357,41 @@ async function bulkReingest(req, res) {
   return res.status(202).send({ id: asyncOperationId });
 }
 
-router.get('/files/get_collection_and_granule_id/:bucket/:key', getFileGranuleAndCollectionByBucketAndKey);
+const bulkArchiveGranulesAsyncWrapperSchema = z.object({
+  batchSize: z.number().optional().default(10000),
+  expirationDays: z.number().optional().default(365),
+});
+const parseBulkArchiveGranulesAsyncWrapperPayload = zodParser('bulkChangeCollection payload', bulkArchiveGranulesAsyncWrapperSchema);
+
+async function bulkArchiveGranulesAsyncWrapper(req, res) {
+  const payload = parseBulkArchiveGranulesAsyncWrapperPayload(req.body);
+  const asyncOperationId = uuidv4();
+  const asyncOperationEvent = {
+    asyncOperationId,
+    callerLambdaName: getFunctionNameFromRequestContext(req),
+    lambdaName: process.env.ArchiveRecordsLambda,
+    description: 'look at me go!',
+    operationType: 'Bulk Granule Reingest',
+    payload: {
+      config: {
+        ...payload,
+        recordType: 'granules',
+      },
+    },
+  };
+  log.debug(
+    `About to invoke lambda to start async operation ${asyncOperationId}`
+  );
+  await startAsyncOperation.invokeStartAsyncOperationLambda(
+    asyncOperationEvent
+  );
+  return res.status(202).send({ id: asyncOperationId });
+}
+
+router.patch('/archiveAsync', bulkArchiveGranulesAsyncWrapper);
+router.patch('/archive', bulkArchiveGranules);
 router.get('/:collectionId/:granuleId', get);
+router.get('/files/get_collection_and_granule_id/:bucket/:key', getFileGranuleAndCollectionByBucketAndKey);
 router.get('/:granuleId', getByGranuleId);
 router.get('/', list);
 router.patch('/bulkPatchGranuleCollection', bulkPatchGranuleCollection);
@@ -1374,6 +1436,7 @@ module.exports = {
   patchGranuleAndReturnStatus,
   bulkPatch,
   bulkPatchGranuleCollection,
+  bulkArchiveGranules,
   getFileGranuleAndCollectionByBucketAndKey,
   router,
 };